use data index for overall ordinal (#783)
this is in prep for a later change that entirely removes the overall
ordinal. It is kept for now but all it is used for is determining
whether to show the pointers

I did just do this for both ordinals to avoid more complicated logic. On
my laptop it consistently adds <1ms (two Date.now() calls do not record
any timing difference at all), so it's fine imo to just do it even if
its a bit redundant in some cases

Adds a bit of delay because otherwise the test was flaking. Not sure why
this just came up (as described above there isn't really any reason this
should be slowing things down?)