Add fullscreen control to maps (#1174)
Adds a full screen button to all maps (in the top left so that it
doesn't interfere with the "Zoom to" control in compare, though it might
seem a bit odd that those have different heights/margins). Maybe worth
asking if this is necessary for the small juxtastat maps but I don't
think it's too cluttery.

Resolves #1170.

I also had to explicitly tell rspack not to watch node_modules in order
to stay under the file watchers limit, though apparently this should be
default behavior.

---------

Co-authored-by: Luke Brody <luke@lukebrody.com>